requirements.txt
beautifulsoup4==4.12.2
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
openai==0.27.8
pandas==2.0.3
python-dotenv==1.0.0
tiktoken==0.4.0
tqdm==4.66.1